+++ backend/app/app/api/api_v1/endpoints/root.py
@@ -63,272 +63,6 @@ def root() -> HTMLResponse:
     return HTMLResponse(html)
 
 
-@router.post("/users/login", response_model=schemas.Token)
-def user_kakao(*, db: Session = Depends(deps.get_db), code: schemas.Code) -> Any:
-
-    get_access_token_headers = {
-        "Content-type": "application/x-www-form-urlencoded;charset=utf-8"
-    }
-    params = {
-        "grant_type": "authorization_code",
-        "client_id": settings.KAKAO_APP_KEY,
-        "redirect_uri": code.redirect_uri,
-        "code": code.code,
-    }
-
-    kakao_access_token = requests.post(
-        "https://kauth.kakao.com/oauth/token",
-        headers=get_access_token_headers,
-        params=params,
-    ).json()
-    if not kakao_access_token.get("access_token"):
-        raise HTTPException(status_code=401, detail="Incorrect code")
-
-    get_user_info_headers = {
-        "Content-type": "application/x-www-form-urlencoded;charset=utf-8",
-        "Authorization": "Bearer " + kakao_access_token["access_token"],
-    }
-
-    kakao_user = requests.post(
-        "https://kapi.kakao.com/v2/user/me", headers=get_user_info_headers
-    ).json()
-
-    if not kakao_user.get("id"):
-        raise HTTPException(status_code=500, detail="Get kakao id error")
-
-    user = crud.user.get_by_kakao_id(db, kakao_id=kakao_user.get("id"))
-
-    if user:
-        access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
-
-        access_token = schemas.Token(
-            access_token=security.create_access_token(
-                user.id, expires_delta=access_token_expires
-            ),
-            token_type="bearer",
-        )
-
-        return access_token
-
-    created_user = crud.user.create_by_kakao_id(
-        db,
-        kakao_id=kakao_user.get("id"),
-        nickname=kakao_user.get("properties").get("nickname"),
-    )
-    access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
-
-    access_token = schemas.Token(
-        access_token=security.create_access_token(
-            created_user.id, expires_delta=access_token_expires
-        ),
-        token_type="bearer",
-    )
-
-    return access_token
-
-
-@router.get("/users/me", response_model=schemas.User)
-def get_user_me(
-    db: Session = Depends(deps.get_db),
-    current_user: schemas.User = Depends(deps.get_current_user),
-) -> Any:
-    """
-    본인 유저 정보 조회
-
-    JWT 필요
-    """
-    return current_user
-
-
-@router.patch("/users/me", response_model=schemas.User)
-def update_user_me(
-    *,
-    db: Session = Depends(deps.get_db),
-    current_user: schemas.User = Depends(deps.get_current_user),
-    user_in: schemas.UserUpdate,
-) -> Any:
-    """
-    본인 유저 정보 수정
-
-    JWT 필요
-    """
-
-    if user_in.nickname is not None and user_in.nickname == "":
-        raise HTTPException(status_code=400, detail="Nickname must not be empty string")
-
-    user_db = crud.user.get(db, id=current_user.id)
-
-    if user_db is None:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    updated_user = crud.user.update(db, db_obj=user_db, obj_in=user_in)
-
-    return updated_user
-
-
-@router.get("/timetables/{timetable_id}", response_model=schemas.TimeTable)
-def get_timetable_by_id(timetable_id: str, db: Session = Depends(deps.get_db)) -> Any:
-    """
-    타임테이블 정보 보기
-    """
-    timetables = crud.timetable.get(db, id=timetable_id)
-    if timetables is None:
-        raise HTTPException(status_code=404, detail="Timetable not found")
-    return timetables
-
-
-@router.get(
-    "/timetables/{timetable_id}/scheduleblocks",
-    response_model=List[schemas.ScheduleBlock],
-)
-def get_scheduleblocks_by_timetable_id(
-    timetable_id: str, db: Session = Depends(deps.get_db)
-) -> Any:
-    """
-    타임테이블의 스케쥴 블록 조회
-    """
-    scheduleblock = crud.scheduleblock.get_all(db, table_id=timetable_id)
-    return scheduleblock
-
-
-@router.get(
-    "/timetables/{timetable_id}/scheduleblocks/me",
-    response_model=List[schemas.ScheduleBlock],
-)
-def get_my_scheduleblocks_by_timetable_id(
-    timetable_id: str,
-    db: Session = Depends(deps.get_db),
-    current_user: schemas.User = Depends(deps.get_current_user),
-) -> Any:
-    """
-    로그인한 유저 본인의 타임테이블의 스케쥴 블록 조회
-    """
-    scheduleblocks = crud.scheduleblock.get_all_by_user_id(
-        db, table_id=timetable_id, user_id=current_user.id
-    )
-    return scheduleblocks
-
-
-@router.post("/timetables", response_model=schemas.TimeTable, status_code=201)
-def create_timetable(
-    *,
-    db: Session = Depends(deps.get_db),
-    timetable_in: schemas.TimeTableCreate,
-    current_user: schemas.User = Depends(deps.get_current_user),
-) -> Any:
-    """
-    타임테이블 생성
-    """
-    timetable = crud.timetable.create_with_user_id(
-        db, obj_in=timetable_in, user_id=current_user.id
-    )
-    if not timetable:
-        raise HTTPException(status_code=500, detail="Timetable is not created")
-    return timetable
-
-
-@router.patch(
-    "/timetables/{timetable_id}", response_model=schemas.TimeTable, status_code=201
-)
-def update_timetable_by_id(
-    *,
-    db: Session = Depends(deps.get_db),
-    timetable_id: str,
-    timetable_in: schemas.TimeTableUpdate,
-    current_user: models.User = Depends(deps.get_current_user),
-) -> Any:
-    """
-    시간표 정보 수정
-    JWT 필요
-    """
-    timetable_db = crud.timetable.get(db, timetable_id)
-
-    if timetable_db is None:
-        raise HTTPException(status_code=404, detail="Timetable not found")
-
-    if timetable_db.create_user_id != current_user.id:
-        raise HTTPException(
-            status_code=403,
-            detail="The user doesn't have enough privileges. Only timetable owner can update info.",
-        )
-
-    return_timetable = crud.timetable.update(
-        db, db_obj=timetable_db, obj_in=timetable_in
-    )
-    return return_timetable
-
-
-@router.post("/scheduleblocks", status_code=201, response_model=schemas.ScheduleBlock)
-def create_scheduleblock(
-    *,
-    db: Session = Depends(deps.get_db),
-    scheduleblock_in: schemas.ScheduleBlockCreate,
-    current_user: models.User = Depends(deps.get_current_user),
-) -> Any:
-    """
-    스케쥴 블록 생성 API
-    """
-    scheduleblock = crud.scheduleblock.create_with_user_id(
-        db, obj_in=scheduleblock_in, user_id=current_user.id
-    )
-
-    return scheduleblock
-
-
-@router.patch("/scheduleblocks", status_code=201)
-def update_scheduleblock_by_id(
-    *,
-    db: Session = Depends(deps.get_db),
-    scheduleblocks_in: List[schemas.ScheduleBlockUpdate],
-    current_user: schemas.User = Depends(deps.get_current_user),
-) -> Any:
-    """ 스케쥴 블록 수정
-    JWT 필요
-    """
-
-    scheduleblocks: List[scheduleblock.ScheduleBlock] = []
-
-    for scheduleblock_in in scheduleblocks_in:
-        scheduleblock_db = crud.scheduleblock.get(db, id=scheduleblock_in.id)
-
-        if scheduleblock_db is None:
-            raise HTTPException(status_code=404, detail="Scheduleblock not found")
-
-        if scheduleblock_db.user_id != current_user.id:
-            raise HTTPException(
-                status_code=403, detail="The user doesn't have enough privileges"
-            )
-
-        updated_scheduleblock = crud.scheduleblock.update(
-            db, db_obj=scheduleblock_db, obj_in=scheduleblock_in
-        )
-
-        scheduleblocks.append(updated_scheduleblock)
-    return scheduleblocks
-
-
-@router.delete("/scheduleblocks/{scheduleblock_id}")
-def delete_scheduleblock_by_id(
-    scheduleblock_id: str,
-    db: Session = Depends(deps.get_db),
-    *,
-    current_user: schemas.User = Depends(deps.get_current_user),
-) -> Any:
-    scheduleblock_db = crud.scheduleblock.get(db, id=scheduleblock_id)
-
-    if scheduleblock_db is None:
-        raise HTTPException(status_code=404, detail="Scheduleblock not found")
-
-    if scheduleblock_db.user_id != current_user.id:
-        raise HTTPException(
-            status_code=403, detail="The user doesn't have enough privileges"
-        )
-
-    _return = crud.scheduleblock.remove(db, id=scheduleblock_db.id)
-
-    return _return
-
-
 class ConnectionManager:
     def __init__(self) -> None:
         self.connections: List[WebSocket] = []
